Split networkControllerPolicyEventHandler into networkControllerPolicyEventHandler and addressSetEventHandler. Now networkControllerPolicyEventHandler uses NetworkPolicyController instead of BaseNetworkController.
