+++ go-controller/pkg/ovn/pod_selector_address_set.go
@@ -3,6 +3,7 @@ package ovn
 import (
 	"fmt"
 	"net"
+	"reflect"
 	"sort"
 	"strings"
 	"sync"
@@ -16,6 +17,7 @@ import (
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/metrics"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
 	addressset "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/address_set"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/retry"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 
 	kapi "k8s.io/api/core/v1"
@@ -24,6 +26,7 @@ import (
 	"k8s.io/apimachinery/pkg/labels"
 	kerrorsutil "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog/v2"
 )
 
@@ -233,7 +236,7 @@ func (bnc *BaseNetworkController) addPodSelectorHandler(psAddrSet *PodSelectorAd
 		_ = bnc.handlePodAddUpdate(podHandlerResources, objs...)
 		return nil
 	}
-	retryFramework := bnc.newNetpolRetryFramework(
+	retryFramework := bnc.newAddressSetRetryFramework(
 		factory.AddressSetPodSelectorType,
 		syncFunc,
 		podHandlerResources)
@@ -254,7 +257,7 @@ func (bnc *BaseNetworkController) addNamespacedPodSelectorHandler(psAddrSet *Pod
 	// start watching namespaces selected by the namespace selector nsSel;
 	// upon namespace add event, start watching pods in that namespace selected
 	// by the label selector podSel
-	retryFramework := bnc.newNetpolRetryFramework(
+	retryFramework := bnc.newAddressSetRetryFramework(
 		factory.AddressSetNamespaceAndPodSelectorType,
 		nil,
 		psAddrSet.handlerResources,
@@ -531,7 +534,7 @@ func (bnc *BaseNetworkController) handleNamespaceAddUpdate(podHandlerInfo *PodSe
 		_ = bnc.handlePodAddUpdate(podHandlerInfo, objs...)
 		return nil
 	}
-	retryFramework := bnc.newNetpolRetryFramework(
+	retryFramework := bnc.newAddressSetRetryFramework(
 		factory.AddressSetPodSelectorType,
 		syncFunc,
 		podHandlerInfo,
@@ -738,4 +741,170 @@ func deleteAddrSetsWithoutACLRef(predicateIDs *libovsdbops.DbObjectIDs,
 		return fmt.Errorf("faile to trasact db ops: %v", err)
 	}
 	return nil
+}
+
+func (bnc *BaseNetworkController) newAddressSetRetryFramework(
+	objectType reflect.Type,
+	syncFunc func([]interface{}) error,
+	extraParameters interface{}) *retry.RetryFramework {
+	eventHandler := &addressSetEventHandler{
+		objType:         objectType,
+		watchFactory:    bnc.watchFactory,
+		bnc:             bnc,
+		extraParameters: extraParameters, // in use by network policy dynamic watchers
+		syncFunc:        syncFunc,
+	}
+	resourceHandler := &retry.ResourceHandler{
+		HasUpdateFunc:          hasResourceAnUpdateFunc(objectType),
+		NeedsUpdateDuringRetry: needsUpdateDuringRetry(objectType),
+		ObjType:                objectType,
+		EventHandler:           eventHandler,
+	}
+	return retry.NewRetryFramework(
+		bnc.stopChan,
+		bnc.wg,
+		bnc.watchFactory,
+		resourceHandler,
+	)
+}
+
+// event handlers handles address set related events
+type addressSetEventHandler struct {
+	retry.EmptyEventHandler
+	watchFactory    *factory.WatchFactory
+	objType         reflect.Type
+	bnc             *BaseNetworkController
+	extraParameters interface{}
+	syncFunc        func([]interface{}) error
+}
+
+// AreResourcesEqual returns true if, given two objects of a known resource type, the update logic for this resource
+// type considers them equal and therefore no update is needed. It returns false when the two objects are not considered
+// equal and an update needs be executed. This is regardless of how the update is carried out (whether with a dedicated update
+// function or with a delete on the old obj followed by an add on the new obj).
+func (h *addressSetEventHandler) AreResourcesEqual(obj1, obj2 interface{}) (bool, error) {
+	// switch based on type
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		// For these types, there was no old vs new obj comparison in the original update code,
+		// so pretend they're always different so that the update code gets executed
+		return false, nil
+
+	case factory.AddressSetNamespaceAndPodSelectorType:
+		// For these types there is no update code, so pretend old and new
+		// objs are always equivalent and stop processing the update event.
+		return true, nil
+	}
+
+	return false, fmt.Errorf("no object comparison for type %s", h.objType)
+}
+
+// GetResourceFromInformerCache returns the latest state of the object, given an object key and its type.
+// from the informers cache.
+func (h *addressSetEventHandler) GetResourceFromInformerCache(key string) (interface{}, error) {
+	var obj interface{}
+	var namespace, name string
+	var err error
+
+	namespace, name, err = cache.SplitMetaNamespaceKey(key)
+	if err != nil {
+		return nil, fmt.Errorf("failed to split key %s: %v", key, err)
+	}
+
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		obj, err = h.watchFactory.GetPod(namespace, name)
+
+	case factory.AddressSetNamespaceAndPodSelectorType:
+		obj, err = h.watchFactory.GetNamespace(name)
+
+	default:
+		err = fmt.Errorf("object type %s not supported, cannot retrieve it from informers cache",
+			h.objType)
+	}
+	return obj, err
+}
+
+// AddResource adds the specified object to the cluster according to its type and returns the error,
+// if any, yielded during object creation.
+// Given an object to add and a boolean specifying if the function was executed from iterateRetryResources
+func (h *addressSetEventHandler) AddResource(obj interface{}, fromRetryLoop bool) error {
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		peerAS := h.extraParameters.(*PodSelectorAddrSetHandlerInfo)
+		return h.bnc.handlePodAddUpdate(peerAS, obj)
+
+	case factory.AddressSetNamespaceAndPodSelectorType:
+		peerAS := h.extraParameters.(*PodSelectorAddrSetHandlerInfo)
+		return h.bnc.handleNamespaceAddUpdate(peerAS, obj)
+
+	default:
+		return fmt.Errorf("no add function for object type %s", h.objType)
+	}
+}
+
+// UpdateResource updates the specified object in the cluster to its version in newObj according to its
+// type and returns the error, if any, yielded during the object update.
+// Given an old and a new object; The inRetryCache boolean argument is to indicate if the given resource
+// is in the retryCache or not.
+func (h *addressSetEventHandler) UpdateResource(oldObj, newObj interface{}, inRetryCache bool) error {
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		peerAS := h.extraParameters.(*PodSelectorAddrSetHandlerInfo)
+		return h.bnc.handlePodAddUpdate(peerAS, newObj)
+	}
+	return fmt.Errorf("no update function for object type %s", h.objType)
+}
+
+// DeleteResource deletes the object from the cluster according to the delete logic of its resource type.
+// Given an object and optionally a cachedObj; cachedObj is the internal cache entry for this object,
+// used for now for pods and network policies.
+func (h *addressSetEventHandler) DeleteResource(obj, cachedObj interface{}) error {
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		peerAS := h.extraParameters.(*PodSelectorAddrSetHandlerInfo)
+		return h.bnc.handlePodDelete(peerAS, obj)
+
+	case factory.AddressSetNamespaceAndPodSelectorType:
+		peerAS := h.extraParameters.(*PodSelectorAddrSetHandlerInfo)
+		return h.bnc.handleNamespaceDel(peerAS, obj)
+
+	default:
+		return fmt.Errorf("object type %s not supported", h.objType)
+	}
+}
+
+func (h *addressSetEventHandler) SyncFunc(objs []interface{}) error {
+	var syncFunc func([]interface{}) error
+
+	if h.syncFunc != nil {
+		// syncFunc was provided explicitly
+		syncFunc = h.syncFunc
+	} else {
+		switch h.objType {
+		case factory.AddressSetNamespaceAndPodSelectorType,
+			factory.AddressSetPodSelectorType:
+			syncFunc = nil
+
+		default:
+			return fmt.Errorf("no sync function for object type %s", h.objType)
+		}
+	}
+	if syncFunc == nil {
+		return nil
+	}
+	return syncFunc(objs)
+}
+
+// IsObjectInTerminalState returns true if the given object is a in terminal state.
+// This is used now for pods that are either in a PodSucceeded or in a PodFailed state.
+func (h *addressSetEventHandler) IsObjectInTerminalState(obj interface{}) bool {
+	switch h.objType {
+	case factory.AddressSetPodSelectorType:
+		pod := obj.(*kapi.Pod)
+		return util.PodCompleted(pod)
+
+	default:
+		return false
+	}
 }
